pyproject.toml
[tool.poetry]
name = "hands-on-deep-research"
version = "0.1.0"
description = "A deep research project inspired by Hugging Face's open-deep-research approach"
authors = ["xiaoye_hua <huag@kth.se>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/xiaoye_hua/hands-on-deep-research"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.26.0"
pandas = "^2.1.0"
torch = "^2.1.0"
transformers = "^4.36.0"
datasets = "^2.14.0"
scikit-learn = "^1.3.0"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
jupyterlab = "^4.0.0"
tqdm = "^4.66.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.0"
python-dotenv = "^1.0.0"
nltk = "^3.9.1"
openai = "^1.65.4"
aiohttp = "^3.11.13"
bs4 = "^0.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.0"
pylint = "^2.17.0"
pre-commit = "^3.3.0"
pytest-cov = "^4.1.0"
sphinx = "^7.1.0"
sphinx-rtd-theme = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
